Refactor CLI project and template handlers
